+++ packages/@lou.codes/ansi/src/controlSequenceIntroducer.ts
@@ -1,4 +1,5 @@
-import type { First, Maybe, Second } from "@lou.codes/types";
+import { EMPTY_STRING } from "@lou.codes/constants";
+import type { EmptyString, First, Maybe, Second } from "@lou.codes/types";
 import { escapeSequence } from "./escapeSequence.js";
 
 /**
@@ -42,12 +43,15 @@ export const controlSequenceIntroducer =
 		escapeSequence(
 			`[${
 				typeof input === "number" ? input : (
-					(`${input[0] ?? ""};${input[1] ?? ""}` as const)
+					(`${input[0] ?? EMPTY_STRING};${
+						input[1] ?? EMPTY_STRING
+					}` as const)
 				)
 			}${code}` as `[${Input extends (
 				readonly [open: Maybe<number>, close: Maybe<number>]
 			) ?
 				`${First<Input> extends number ? First<Input>
-				:	""};${Second<Input> extends number ? Second<Input> : ""}`
+				:	EmptyString};${Second<Input> extends number ? Second<Input>
+				:	EmptyString}`
 			:	`${Input & number}`}${Code}`,
 		);