+++ packages/@lou.codes/ansi/src/optionalFormat.ts
@@ -1,3 +1,5 @@
+import { EMPTY_OBJECT } from "@lou.codes/constants";
+import { EMPTY_STRING } from "@lou.codes/constants/EMPTY_STRING.js";
 import type { Maybe } from "@lou.codes/types";
 import { format } from "./format.js";
 import { normalizeString } from "./normalizeString.js";
@@ -24,15 +26,20 @@ import { normalizeString } from "./normalizeString.js";
  * @param process NodeJS `globalThis.process`.
  * @returns Either the formatted string, or just the passed string.
  */
-export const optionalFormat = (({ process: { env = {}, stdout = {} } = {} }) =>
+export const optionalFormat = (({
+	process: {
+		env = EMPTY_OBJECT as Partial<(typeof globalThis)["process"]["env"]>,
+		stdout = { isTTY: false },
+	} = EMPTY_OBJECT as Partial<(typeof globalThis)["process"]>,
+}) =>
 	(
-		(env.NODE_DISABLE_COLORS ?? "") === "" &&
+		(env.NODE_DISABLE_COLORS ?? EMPTY_STRING) === EMPTY_STRING &&
 		env.NO_COLOR === undefined &&
 		env.TERM !== "dumb" &&
-		((env.FORCE_COLOR ?? "1") !== "0" || (stdout.isTTY ?? false))
+		((env.FORCE_COLOR ?? "1") !== "0" || stdout.isTTY)
 	) ?
 		format
 	:	() => () => normalizeString)(
 	(globalThis as Maybe<Partial<typeof globalThis>>) ??
-		({} as Partial<typeof globalThis>),
+		(EMPTY_OBJECT as Partial<typeof globalThis>),
 ) as typeof format;