+++ packages/@lou.codes/constants/src/EMPTY_ARRAY.ts
@@ -1,0 +1,4 @@
+/**
+ * Empty read only array.
+ */
+export const EMPTY_ARRAY = Object.freeze([]) as readonly [];