+++ packages/@lou.codes/constants/src/EMPTY_OBJECT.ts
@@ -1,0 +1,6 @@
+import { createObject } from "./createObject.js";
+
+/**
+ * Empty read only object.
+ */
+export const EMPTY_OBJECT = Object.freeze(createObject());