+++ packages/@lou.codes/constants/src/NIL.ts
@@ -1,0 +1,4 @@
+/**
+ * Alias for `undefined`
+ */
+export const NIL = undefined;