+++ packages/@lou.codes/constants/src/createObject.ts
@@ -1,0 +1,7 @@
+/**
+ * Alias for `Object.create(null)`.
+ *
+ * @returns Object with no prototype.
+ */
+// eslint-disable-next-line no-null/no-null, @typescript-eslint/ban-types
+export const createObject = () => Object.create(null) as {};