+++ packages/@lou.codes/create-package/src/userClean.ts
@@ -1,3 +1,5 @@
+import { EMPTY_STRING } from "@lou.codes/constants";
 import { USER } from "./constants.js";
 
-export const userClean = (name: string) => name.replace(`@${USER}/`, "");
+export const userClean = (name: string) =>
+	name.replace(`@${USER}/`, EMPTY_STRING);