+++ packages/@lou.codes/create-package/src/validations/startsWithPeriod.ts
@@ -1,6 +1,7 @@
 import { foregroundRed } from "@lou.codes/ansi";
+import { EMPTY_STRING } from "@lou.codes/constants";
 
 export const startsWithPeriod = (value: string) =>
 	value.startsWith(".") ?
 		foregroundRed`"${value}" must not start with a period`
-	:	"";
+	:	EMPTY_STRING;