+++ packages/@lou.codes/diff/tests/test.ts
@@ -1,3 +1,4 @@
+import { EMPTY_STRING } from "@lou.codes/constants";
 import type { Maybe } from "@lou.codes/types";
 import { deepEqual } from "node:assert/strict";
 import { describe, it } from "node:test";
@@ -7,7 +8,7 @@ import type { Tests } from "./types/Tests.js";
 export const test = (fileURL: string) => {
 	const relativePath = fileURLToPath(fileURL).replace(
 		(globalThis.process as Maybe<(typeof globalThis)["process"]>)?.cwd() ??
-			"",
+			EMPTY_STRING,
 		".",
 	);