+++ packages/@lou.codes/test/src/stringifyDifferenceDictionary.ts
@@ -1,4 +1,5 @@
 import { foregroundRed } from "@lou.codes/ansi";
+import { EMPTY_ARRAY } from "@lou.codes/constants";
 import { CREATE, DELETE, UPDATE } from "@lou.codes/diff";
 import { EXCEPTION } from "./constants.js";
 import { formatPropertyPath } from "./formatPropertyPath.js";
@@ -15,16 +16,17 @@ export const stringifyDifferenceDictionary: {
 		difference: Difference & { readonly kind: Kind },
 	) => string;
 } = {
-	[CREATE]: ({ path = [], right: received }) =>
+	[CREATE]: ({ path = EMPTY_ARRAY, right: received }) =>
 		`${formatPropertyPath(path)} was set with value ${formatValue(
 			received,
 		)}.`,
-	[DELETE]: ({ path = [] }) => `${formatPropertyPath(path)} is missing.`,
+	[DELETE]: ({ path = EMPTY_ARRAY }) =>
+		`${formatPropertyPath(path)} is missing.`,
 	[EXCEPTION]: ({ error }) =>
 		foregroundRed`there was an uncaught error: ${
 			error instanceof Error ? error.message : (error as string)
 		}.`,
-	[UPDATE]: ({ path = [], right: received, left: wanted }) =>
+	[UPDATE]: ({ path = EMPTY_ARRAY, right: received, left: wanted }) =>
 		`${formatPropertyPath(path)} has the wrong value. Wanted ${formatValue(
 			wanted,
 		)} but received ${formatValue(received)}.`,