+++ packages/@lou.codes/create-package/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## 1.0.3
+
+### Patch Changes
+
+-   ➕ add @lou.codes/constants and use it.
+-   ⬆️ upgrade `@lou.codes/prompts` to `v1.0.2`
+-   ⬆️ upgrade `@lou.codes/types` to `v1.0.2`
+-   ⬆️ upgrade `@lou.codes/ansi` to `v1.0.2`
+
 ## 1.0.2
 
 ### Patch Changes